Add portable interface to use binary mode stdin
Currently, all platform specific code is declared in xplat.hpp and
implemented in files in src/xplat/. The files in src/xplat/ use
preprocessor conditionals to create an implementation section for each
platform.

There may be a better way to organize this but this structure should be
good enough for a small scale project.